BUG#: 5185 TITLE: JMPI Windows build break
DESCRIPTION:
Fixed Windows build break under JMPI

+++ src/Pegasus/ProviderManager2/JMPI/org/pegasus/jmpi/tests/Providers/Indications/TestIndicationsListener.cpp
@@ -38,7 +38,7 @@ public:
             << " ! "
             << indicationInstance.getClassName ().getString ()
             << " ! "
-            << ui64Number
+            << (int)ui64Number
             << endl;
     }
